requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
cycler==0.12.1
DateTime==5.5
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
executing==2.2.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
fastjsonschema==2.21.1
filelock==3.19.1
Flask==3.1.1
Flask-Mail==0.10.0
fonttools==4.60.0
fqdn==1.5.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
narwhals==1.41.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.2.6
opencv-python==4.12.0.88
orjson==3.11.3
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
polars==1.33.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.25.1
scipy==1.16.2
Send2Trash==1.8.3
sentry-sdk==2.38.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.48.0
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
tinycss2==1.4.0
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.1
traitlets==5.14.3
typer==0.19.2
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.11.0
ultralytics==8.3.203
ultralytics-thop==2.0.17
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.37.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
zope.interface==7.2
